Polish code
