+++ src/main/java/com/hazelcast/eureka/one/EurekaHazelcastMetadata.java
@@ -3,9 +3,9 @@ package com.hazelcast.eureka.one;
 /**
  * Hazelcast node address and port provided in Eureka metadata map.
  */
-public class EurekaHazelcastMetadata {
+final class EurekaHazelcastMetadata {
 
-    public static final String HAZELCAST_HOST = "hazelcast.host";
-    public static final String HAZELCAST_PORT = "hazelcast.port";
-    public static final String HAZELCAST_GROUP_NAME = "hazelcast.groupName";
+    static final String HAZELCAST_HOST = "hazelcast.host";
+    static final String HAZELCAST_PORT = "hazelcast.port";
+    static final String HAZELCAST_GROUP_NAME = "hazelcast.groupName";
 }
